make check of creation of library a true/false for forntend, add security in api call, clean tests
